mian_process/requirements.txt
# 影像處理相關
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.21.0
matplotlib>=3.5.0

# 機器學習/深度學習
torch>=1.13.0
torchvision>=0.14.0
ultralytics>=8.0.0  # YOLO v8

# 音訊處理
librosa>=0.9.0
pydub>=0.25.0
demucs>=4.0.0

# 影片處理
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# 資料處理
pandas>=1.5.0
pyyaml>=6.0
tqdm>=4.64.0

# API相關 (LLM整合)
openai>=1.0.0
requests>=2.28.0

# 開發工具
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0